Move snapshot-specific checks into SnapshotMetadata
