+++ src/Metadata/MetaFileInfoTrait.php
@@ -23,29 +23,8 @@ trait MetaFileInfoTrait
                         "is less than previously seen  version \"${$localFileInfo['version']}\"";
                     throw new RollbackAttackException($message);
                 }
-            } elseif ($type === 'snapshot' && static::getFileNameType($fileName) === 'targets') {
-                // TUF-SPEC-v1.0.16 Section 5.4.4
-                // Any targets metadata filename that was listed in the trusted snapshot metadata file, if any, MUST
-                // continue to be listed in the new snapshot metadata file.
-                throw new RollbackAttackException("Remote snapshot metadata file references '$fileName' but this is not present in the remote file");
             }
         }
-    }
-
-    /**
-     * Gets the type for the file name.
-     *
-     * @param string $fileName
-     *   The file name.
-     *
-     * @return string
-     *   The type.
-     */
-    private static function getFileNameType(string $fileName): string
-    {
-        $parts = explode('.', $fileName);
-        array_pop($parts);
-        return array_pop($parts);
     }
 
     /**
