+++ src/Metadata/SnapshotMetadata.php
@@ -7,6 +7,7 @@ use Symfony\Component\Validator\Constraints\Count;
 use Symfony\Component\Validator\Constraints\Required;
 use Symfony\Component\Validator\Constraints\Type;
 use Tuf\Constraints\Collection;
+use Tuf\Exception\PotentialAttackException\RollbackAttackException;
 
 class SnapshotMetadata extends MetadataBase
 {
@@ -23,6 +24,35 @@ class SnapshotMetadata extends MetadataBase
     {
         parent::checkRollbackAttack($remoteMetadata, $expectedRemoteVersion);
         $this->traitCheckRollbackAttack($remoteMetadata, $expectedRemoteVersion);
+
+        $localMetadata = $this;
+        $localMetaFileInfos = $localMetadata->getSigned()['meta'];
+        foreach ($localMetaFileInfos as $fileName => $localFileInfo) {
+            /** @var \Tuf\Metadata\SnapshotMetadata|\Tuf\Metadata\TimestampMetadata $remoteMetadata */
+            $remoteFileInfo = $remoteMetadata->getFileMetaInfo($fileName, true);
+            if (empty($remoteFileInfo) && static::getFileNameType($fileName) === 'targets') {
+                // TUF-SPEC-v1.0.16 Section 5.4.4
+                // Any targets metadata filename that was listed in the trusted snapshot metadata file, if any, MUST
+                // continue to be listed in the new snapshot metadata file.
+                throw new RollbackAttackException("Remote snapshot metadata file references '$fileName' but this is not present in the remote file");
+            }
+        }
+    }
+
+    /**
+     * Gets the type for the file name.
+     *
+     * @param string $fileName
+     *   The file name.
+     *
+     * @return string
+     *   The type.
+     */
+    private static function getFileNameType(string $fileName): string
+    {
+        $parts = explode('.', $fileName);
+        array_pop($parts);
+        return array_pop($parts);
     }
 
     /**
